+++ backend/controllers/product.js
@@ -1,6 +1,5 @@
 const { Product } = require('../models')
 const { body, validationResult } = require('express-validator');
-const { v4: uuidv4 } = require('uuid');
 
 /**
  * Creates a new product
@@ -8,19 +7,32 @@ const { v4: uuidv4 } = require('uuid');
  * @param {*} res
  * @returns Object
  */
-const createProduct = async (req, res) => {
-  try {
-    const data = { ...req.body, id: uuidv4(), quantity: parseInt(req.body.quantity), price: parseInt(req.body.price), createdAt: new Date(), updatedAt: new Date() }
+const createProduct = [
+  body('name').notEmpty().withMessage('Nome é obrigatório'),
+  body('price').isNumeric().withMessage('O preço deve ser numérico'),
 
-    const product = await Product.create(data)
+  async (req, res) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      return res.status(400).json({ errors: errors.array() });
+    }
 
-    return res.status(201).json(
-      product
-    )
-  } catch (error) {
-    return res.status(500).json({ error: error.message })
+    // Transformação dos dados
+    const transformedData = {
+      ...req.body,
+      name: req.body.name.toLowerCase()
+    };
+
+    try {
+      const product = await Product.create(transformedData);
+      return res.status(201).json(
+        product
+      );
+    } catch (error) {
+      return res.status(500).json({ error: error.message });
+    }
   }
-}
+];
 
 
 /**
@@ -67,22 +79,37 @@ const getProductById = async (req, res) => {
  * @param {*} res
  * @returns boolean
  */
-const updateProductById = async (req, res) => {
-  try {
-    const { id } = req.params
-    const product = await Product.update(req.body, {
-      where: { id: id }
-    })
+const updateProductById = [
+  body('name').optional().notEmpty().withMessage('Nome não pode estar vazio'),
+  body('price').optional().isNumeric().withMessage('O preço deve ser numérico'),
 
-    if (product) {
-      const updatedProduct = await Product.findOne({ where: { id: id } })
-      return res.status(200).json(updatedProduct)
+  async (req, res) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      return res.status(400).json({ errors: errors.array() });
     }
-    throw new Error('product not found')
-  } catch (error) {
-    return res.status(500).send(error.message)
+
+    try {
+      const { id } = req.params;
+      let product = await Product.findOne({ where: { id: id } });
+
+      if (!product) {
+        return res.status(404).send('Product not found');
+      }
+
+      // Transformação de dados antes de atualizar
+      const updatedData = req.body;
+      if (updatedData.name) {
+        updatedData.name = updatedData.name.toLowerCase(); // Converter nome para minúsculo
+      }
+
+      await product.update(updatedData);
+      return res.status(200).json( product );
+    } catch (error) {
+      return res.status(500).send(error.message);
+    }
   }
-}
+];
 
 
 /**
